+++ src/base/ElementExporterInterface.php
@@ -29,8 +29,9 @@ interface ElementExporterInterface extends ComponentInterface
      * Sets the element type on the exporter.
      *
      * @param string $elementType
+     * @return void
      */
-    public function setElementType(string $elementType);
+    public function setElementType(string $elementType): void;
 
     /**
      * Creates the export data for elements fetched with the given element query.